notifier: add Notifier class and sigrt injection
This combo is based on inotifier in order to connect the source code to sac.
To do so, we create a deamon (Notifier) which will poll the inotify events
from the source code. In case of modifications, the Notifier will inject
a SIGRTMIN signal to the tracee. The signal will trigger an handler into sac
and give it back the controle of the debugging session.

Notifier and sac shared a modifications set which is populated by Notifier
and then cleared by sac once it has built and injected the modifications.

Signed-off-by: Alexandre Bernard <alexandre.bernard@lse.epita.fr>